fix: Error when executing last cell without bottom marker
We were getting an error when executing the last cell which has an
implicit cell marker at the bottom. The reason was that we request +1
lines but when we are at the bottom of the script there is no such +1
line and therefore Molten was rightly erroring out.

The fix is to first check and if there is no bottom marker adding one.